+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/service/bulk/print/dataextractor/ServiceApplicationDataExtractor.java
@@ -1,21 +1,18 @@
-package uk.gov.hmcts.reform.divorce.orchestration.util;
+package uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
-import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CaseDetails;
 
 import java.util.Map;
 
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdFields.SERVICE_APPLICATION_GRANTED;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdFields.SERVICE_APPLICATION_REFUSAL_REASON;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdFields.SERVICE_APPLICATION_TYPE;
-import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdStates.AWAITING_SERVICE_CONSIDERATION;
-import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.YES_VALUE;
 import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.helper.ExtractorHelper.getMandatoryStringValue;
 import static uk.gov.hmcts.reform.divorce.orchestration.tasks.util.TaskUtils.getOptionalPropertyValueAsString;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class ServiceApplicationRefusalHelper {
+public class ServiceApplicationDataExtractor {
 
     public static String getServiceApplicationRefusalReason(Map<String, Object> caseData) {
         return getMandatoryStringValue(caseData, SERVICE_APPLICATION_REFUSAL_REASON);
@@ -33,11 +30,4 @@ public class ServiceApplicationRefusalHelper {
         return getMandatoryStringValue(caseData, SERVICE_APPLICATION_TYPE);
     }
 
-    public static boolean isAwaitingServiceConsideration(CaseDetails caseDetails) {
-        return AWAITING_SERVICE_CONSIDERATION.equalsIgnoreCase(caseDetails.getState());
-    }
-
-    public static boolean isServiceApplicationGranted(Map<String, Object> caseData) {
-        return YES_VALUE.equalsIgnoreCase(getServiceApplicationGranted(caseData));
-    }
 }
