+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/tasks/servicejourney/DispensedServiceRefusalOrderDraftTask.java
@@ -15,8 +15,8 @@ import uk.gov.hmcts.reform.divorce.orchestration.util.CcdUtil;
 import java.util.Map;
 
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.document.ApplicationServiceTypes.DISPENSED;
+import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.ServiceApplicationDataExtractor.getServiceApplicationRefusalReason;
 import static uk.gov.hmcts.reform.divorce.orchestration.tasks.util.TaskUtils.getCaseId;
-import static uk.gov.hmcts.reform.divorce.orchestration.util.ServiceApplicationRefusalHelper.getServiceApplicationRefusalReason;
 
 @Component
 public class DispensedServiceRefusalOrderDraftTask extends ServiceRefusalOrderDraftTask {
